% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev2df.R
\name{fev2df}
\alias{fev2df}
\title{Convert a palmscan Field-Equals-Value (FEV) column into a dataframe}
\usage{
fev2df(fev.col)
}
\arguments{
\item{fev.col}{character vector of FEV}
}
\value{
A 1-column data.frame
}
\description{
Convert a palmscan Field-Equals-Value (FEV) column into a dataframe
}
\examples{
fev.df <- as.data.frame( apply(fev.tsv, 2, fev2df) )
}
\keyword{fev}
\keyword{palmid}
