% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoFilter.R
\name{geoFilter}
\alias{geoFilter}
\title{Conversion between run_ids and geo objects often contain NA/NULL values
This removes NA-containing rows}
\usage{
geoFilter(sra.geo)
}
\arguments{
\item{sra.geo}{data.frame, output df from get.sraGeo()}
}
\value{
data.frame
}
\description{
Conversion between run_ids and geo objects often contain NA/NULL values
This removes NA-containing rows
}
\examples{

sra.geo.filtered <- geoFilter(sra.geo.full)

}
\keyword{geo}
\keyword{palmid}
