% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nickname.R
\name{get.nickname}
\alias{get.nickname}
\title{get.nickname}
\usage{
get.nickname(palm_ids, con, ordinal = FALSE)
}
\arguments{
\item{palm_ids}{character, set of 'palm_id' to lookup in palmdb}

\item{con}{pq-connection, use SerratusConnect()}

\item{ordinal}{boolean, return an ordered sOTU vector based on input 'palm_ids' [FALSE]}
}
\value{
character, nickname for each palm_id
}
\description{
Retrieve the nickname for a given 'palm_id' in palmdb
}
\examples{

## R Code Example
\donttest{ 
con <- SerratusConnect()
get.nickname(c("u1337"), con)
}

}
\keyword{Serratus}
\keyword{palmdb}
\keyword{palmid}
\keyword{sOTU}
