% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.palmSra.R
\name{get.palmSra}
\alias{get.palmSra}
\title{A wrapper of several get* functions to create a palm.sra data.frame}
\usage{
get.palmSra(pro.df, con = SerratusConnect())
}
\arguments{
\item{pro.df}{data.frame, imported diamond pro df. use get.pro()}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
palm.sra  data.frame
}
\description{
A wrapper of several get* functions to create a palm.sra data.frame
}
\examples{

data("waxsys.pro.df")
con <- SerratusConnect()

\donttest{
get.palmSra( waxsys.pro.df, con )
}

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{biosample}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
