% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.palmVIrome.R
\name{get.palmVirome}
\alias{get.palmVirome}
\title{A wrapper of several get* functions to create a palm.virome data.frame}
\usage{
get.palmVirome(run.vec = NA, org.search = NA, con = SerratusConnect())
}
\arguments{
\item{run.vec}{character vector, sra run accession list}

\item{org.search}{string, search term to query "scientific_name". Uses SQL "like"}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
palm.virome  data.frame
}
\description{
A wrapper of several get* functions to create a palm.virome data.frame
}
\examples{

con <- SerratusConnect()

\donttest{
get.palmVirome(org.search = 'Saccharomyces\%', con )
}

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{bioproject}
\keyword{biosample}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
\keyword{timeline}
