% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.proPhy.R
\name{get.proPhy}
\alias{get.proPhy}
\title{A wrapper for get.palmTax() specific for 'pro.df' input
and returns a populated the "tspe", "tfam", and "tphy"
columns of 'pro.df' based on the "sseqid" column}
\usage{
get.proPhy(pro.df, tree.phy)
}
\arguments{
\item{pro.df}{data.frame, imported diamond pro df. use get.pro()}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
pro.df  data.frame
}
\description{
A wrapper for get.palmTax() specific for 'pro.df' input
and returns a populated the "tspe", "tfam", and "tphy"
columns of 'pro.df' based on the "sseqid" column
}
\examples{

## Prepare data
# data("waxsys.pro.df")
# con <- SerratusConnect()

## Generate Report
# geoSRA <- PlotGeoReport( waxsys.pro.df )


}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
