% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.proTax.R
\name{get.proTax}
\alias{get.proTax}
\title{A wrapper for get.tax() specific for `pro.df` input
and returns a populated the "tspe", "tfam", and "tphy"
columns of `pro.df` based on the "sseqid" column}
\usage{
get.proTax(pro.df, con = SerratusConnect())
}
\arguments{
\item{pro.df}{data.frame, imported diamond pro df. use get.pro()}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
pro.df  data.frame
}
\description{
A wrapper for get.tax() specific for `pro.df` input
and returns a populated the "tspe", "tfam", and "tphy"
columns of `pro.df` based on the "sseqid" column
}
\examples{

geoSRA <- PlotGeoReport( XXX )

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
