% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sOTU.R
\name{get.sOTU}
\alias{get.sOTU}
\title{get.sOTU}
\usage{
get.sOTU(palm_ids, con, get_childs = FALSE, ordinal = FALSE)
}
\arguments{
\item{palm_ids}{character, set of `palm_id` to lookup in palmdb}

\item{con}{pq-connection, use SerratusConnect()}

\item{get_childs}{boolean, return all children `palm_id` instead of parent sOTU [F]}

\item{ordinal}{boolean, return an ordered sOTU vector based on input `palm_ids`}
}
\value{
character, unique `palm_id` sOTU or sOTU-children
}
\description{
Retrieve the parent sOTU for a `palm_id` in palmdb
or the set of all children `palm_id` within an species/sOTU
}
\keyword{Serratus}
\keyword{palmdb}
\keyword{palmid}
\keyword{sOTU}
