% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraDate.R
\name{get.sraDate}
\alias{get.sraDate}
\title{get.sraDate}
\usage{
get.sraDate(run_ids, con, ordinal = FALSE, as.df = FALSE)
}
\arguments{
\item{run_ids}{character, SRA 'run_id'}

\item{con}{pq-connection, use SerratusConnect()}

\item{ordinal}{boolean, return 'run_ids' ordered vector [F]}

\item{as.df}{boolean, return run_id, date data.frame [F]}
}
\value{
POSIXct, date object vector
}
\description{
Retrieve the "Load_Date" for a set of SRA 'run_id'
}
\examples{
\donttest{
con <- SerratusConnect()
palm.date   <- get.sraDate("SRR9968562", con)
}
}
\keyword{Serratus}
\keyword{palmid}
\keyword{timeline}
