% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraGeo.R
\name{get.sraGeo}
\alias{get.sraGeo}
\title{get.sraGeo}
\usage{
get.sraGeo(run_ids = NULL, biosample_ids = NULL, con, ordinal = FALSE)
}
\arguments{
\item{run_ids}{character, SRA `run_id`}

\item{biosample_ids}{character, BioSample `biosample_id`}

\item{con}{pq-connection, use SerratusConnect()}

\item{ordinal}{boolean, return `run_ids` ordered vector [F]}
}
\value{
data.frame, lon and lat numeric vectors
}
\description{
Retrieve the geo_coordinates for a set of SRA `run_id`
}
\examples{
palm.geo   <- get.sraGeo(palm.sras, con)

}
\keyword{Serratus}
\keyword{geo}
\keyword{palmid}
