% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraSTAT.R
\name{get.sraSTAT}
\alias{get.sraSTAT}
\title{get.sraSTAT}
\usage{
get.sraSTAT(run_ids, con = SerratusConnect(), prc_threshold = 5)
}
\arguments{
\item{run_ids}{character, SRA 'run_id'}

\item{con}{pq-connection, use SerratusConnect()}

\item{prc_threshold}{numeric, threshold percent of total reads for tax to retain [5]}
}
\value{
data.frame, with 'run_id', 'order_name', 'label' columns
}
\description{
Retrieve the STAT-taxonomy (at the rank of 'order') and the grouping labels
for a set of SRA run_ids
}
\details{
Data from NCBI-STAT (Katz et al., 2021)
}
\examples{
\donttest{
 # Retrieve STAT taxonomic orders for a SRA run
 # and show all results with prc_threshold of 0\%
con <- SerratusConnect()
sra.taxa   <- get.sraTax('ERR2756788', con, prc_threshold = 0)

}
}
\keyword{STAT}
\keyword{Serratus}
\keyword{palmid}
\keyword{taxonomy}
