% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identityWordcount.R
\name{identityWordcount}
\alias{identityWordcount}
\title{identityWordcount}
\usage{
identityWordcount(orgn.df, ntop = 50)
}
\arguments{
\item{orgn.df}{data.frame, scientific_name, pident}

\item{ntop}{numeric,   return only N top words [50]}
}
\value{
table,   identity-count table with 100% - 20% standard
}
\description{
Create frequency-count table from a set of characters
which are assigned a standardized rank-order scores
from 10.0 to 1.0.
}
\examples{

}
\keyword{palmid}
\keyword{plot}
\keyword{wordcloud}
