% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkBLAST.R
\name{linkBLAST}
\alias{linkBLAST}
\title{Parse an input sequence into a BLAST-able HTML link}
\usage{
linkBLAST(header, aa.seq, label = "[BLAST]")
}
\arguments{
\item{header}{character, header for blast search}

\item{aa.seq}{character, query sequence (amino acid)}

\item{label}{character, Display string["BLAST"]}
}
\value{
character, html link for click to search
}
\description{
Parse an input sequence into a BLAST-able HTML link
}
\examples{

blast.link <- linkBLAST('u1337', 'MYAASTRING', 'BLAST')

}
\keyword{BLAST}
\keyword{html}
\keyword{palmid}
\keyword{sql}
