% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkDB.R
\name{linkDB}
\alias{linkDB}
\title{Parse an input sequence into an Accession-based HTML link}
\usage{
linkDB(
  accession,
  DB = "sra",
  base_url = "https://www.ncbi.nlm.nih.gov",
  label = "<acc>",
  prefix_text = NULL,
  suffix_text = NULL
)
}
\arguments{
\item{accession}{character, SRA run_id accession to look-up}

\item{DB}{character, GenBank database to prefix ["sra"]}

\item{base_url}{character, URL ["https://www.ncbi.nlm.nih.gov/"]}

\item{label}{character, Display string or accession var ["<acc>"]}

\item{prefix_text}{character, prefix to add to label when using accession}

\item{suffix_text}{character, suffix to add to label when using accession}
}
\value{
character, html link for click to search
}
\description{
Parse an input sequence into an Accession-based HTML link
}
\examples{

# Standard SRA link database
db.link <- linkDB("ERR2756788")
db.link <- linkDB("SRR7287110", label = "Ginger The Cat")

# Link to BioSamples Database
db.link <- linkDB("SRR7287110", label = "Ginger The Cat")

}
\keyword{BioProjects}
\keyword{BioSamples}
\keyword{NCBI}
\keyword{SRA}
\keyword{html}
\keyword{palmid}
\keyword{sql}
