% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGB.R
\name{linkGB}
\alias{linkGB}
\title{Parse an input sequence into a GenBank HTML link}
\usage{
linkGB(accession, label = "<accession>", DB = "protein", prefix_text = NULL)
}
\arguments{
\item{accession}{character, header for blast search}

\item{label}{character, Display string or accesion var ["<accesion>"]}

\item{database}{character, GenBank database to prefix ["protein"]}
}
\value{
character, html link for click to search
}
\description{
Parse an input sequence into a GenBank HTML link
}
\examples{

gb.link <- linkGB("VVX76773.1")
gb.link <- linkGB("NC_045512.2", label = "SARSCOV2", DB = "nuccore")

}
\keyword{BLAST}
\keyword{html}
\keyword{palmid}
\keyword{sql}
