% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lib.R
\name{load.lib}
\alias{load.lib}
\title{load.lib.R}
\usage{
load.lib(lib_set)
}
\arguments{
\item{lib_set}{character, one of "sql", "ggplot", "geo", "geo2"}
}
\value{
loads set of packages for functions
}
\description{
Internal function to load a set of libraries for palmid functions
"geo" is for static maps, "geo2" is for htmlwidget maps
}
\examples{

# in an SQL function
# load.lib("sql")

}
\keyword{palmid}
\keyword{serratus}
\keyword{sql}
