% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt.virome.R
\name{melt.virome}
\alias{melt.virome}
\title{Melt the virome.df object into a summarized form around sOTU}
\usage{
melt.virome(virome.df)
}
\arguments{
\item{virome.df}{data.frame, created by get.palmVirome()}
}
\value{
virx.df  data.frame
}
\description{
Melt the virome.df object into a summarized form around sOTU
}
\examples{


\donttest{
virx.df <- melt.virome(virome.df)
}

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{bioproject}
\keyword{biosample}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
\keyword{timeline}
