% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmscan.df.R
\docType{class}
\name{palmscan.df-class}
\alias{palmscan.df-class}
\alias{palmscan.df}
\title{palmscan.df}
\description{
An S4-Object defining the palmscan data.frame. Read from .fev files
with the read.fev() function.
}
\details{
df.cols <- c(
 "score", "query", "gene",	"order", "confidence",
 "qlen",	"pp_start",	"pp_end",	"pp_length",
 "v1_length", "v2_length",
 "pssm_total_score", "pssm_min_score",
 "motifs",	"super", 'group', 'comments' )
}
\section{Slots}{

\describe{
\item{\code{fev}}{palmscan fev}
}}

\keyword{data.frame}
\keyword{fev}
\keyword{palmid}
\keyword{palmscan}
