% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.cladeogram.R
\name{phyloseq.cladeogram}
\alias{phyloseq.cladeogram}
\title{phyloseq.cladeogram}
\usage{
phyloseq.cladeogram(dat, con = SerratusConnect())
}
\arguments{
\item{dat}{phyloseq object derived from phyloseq.palmid()}

\item{con}{Serratus connection}
}
\value{
phyloseq object with phyloseq-tree slot filled with taxa
}
\description{
Add a cladeogram to a phyloseq object phylo-tree slot object for all Serratus IDs found in input
}
\seealso{
[phyloseq.palmid, SerratusConnect()]
}
