% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.downsample.R
\name{phyloseq.downsample}
\alias{phyloseq.downsample}
\title{phyloseq.downsample}
\usage{
phyloseq.downsample(dat, decile.range = c(0, 0.9))
}
\arguments{
\item{dat}{non-normalized phyloseq object derived from phyloseq.palmid()}

\item{decile.range}{downsample decile range. numeric vector of length 2. default is 0 (min) to 90%}
}
\value{
list:
physeq = list of 11 phyloseq objects downsampled to min/max (n = 2) and each decile (n = 9)
downsample_depths =
}
\description{
Downsample phyloseq object per decile. Phyloseq object should count raw coverage count data (non-normalized)

n viral counts sampled is dependent on the sample library size. For example, if the sample library size = 100 and total viral counts = 10, then when downsampling to 10 counts, only 1 (10/100 * 10 = 1) viral count will be sampled. This is to account for the total library size, and not just the number of viruses identified. 
Proportions are rounded to the nearest whole number (e.g 22 viral counts/100 library size * 10 downsample = 2.2  = 2 counts sampled. 
Currently, any samples that have 0 counts when downsamples are removed.
}
\seealso{
[phyloseq.palmid]
}
