% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.merge.group.R
\name{phyloseq.merge.group}
\alias{phyloseq.merge.group}
\title{phyloseq.merge.group}
\usage{
phyloseq.merge.group(dat, group_by = "bio_sample")
}
\arguments{
\item{dat}{phyloseq object}

\item{group_by}{character string, metadata variable used to collapse samples}
}
\value{
phyloseq object with samples merged by group_by variable
}
\description{
Merge samples by some metadata variable. Metadata will be collapsed and separated by |. 
The only exception is sequencing depth (spots), which will be both summed (spots) and collapse by | (spots_sep)

By default performs merge on bio_samples. Note: it is recommended to merge only after per-run normalization and pre-processing.
}
\seealso{
[phyloseq.normalize(), phyloseq.palmid]
}
