% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.normalize.R
\name{phyloseq.normalize}
\alias{phyloseq.normalize}
\title{phyloseq.normalize}
\usage{
phyloseq.normalize(dat, method = "cpm")
}
\arguments{
\item{dat}{phyloseq object generated from phyloseq.palmid}

\item{method}{Normalization method, character string. Currently only accepts "cpm" (Counts Per Million).}
}
\value{
phyloseq object with normalized coverage
}
\description{
Normalize palm_id coverage "counts". Currently only implements Coverage Per Million (comparable to reads per million). Expects phyloseq object with format and metadata generated by phyloseq.palmid function.
}
\seealso{
[phyloseq.palmid()]
}
