% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.palmid.help.count.R
\name{phyloseq.palmid.help.count}
\alias{phyloseq.palmid.help.count}
\title{phyloseq.palmid.help.query}
\usage{
phyloseq.palmid.help.count(palm, sample = "run_id", taxa = "sotu", threads = 1)
}
\arguments{
\item{palm}{character vector of SRA run IDs to generate phyloseq object.}

\item{sample}{user-defined sample column. Recommend to use either run_id (each run as sample) or Biosample (each biological sample, including replicates, as sample)}

\item{taxa}{one of c("sotu", "palm_id").}
}
\value{
Taxa x sample count matrix. Taxa are rows. Columns are samples.
}
\description{
Takes a dataframe queried from Serratus table palm_sra2 and derives a taxa x sample count matrix. Counts are aggregated from coverage per taxa per sample.
}
