% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.palmid.help.query.R
\name{phyloseq.palmid.help.query}
\alias{phyloseq.palmid.help.query}
\title{phyloseq.palmid.help.query}
\usage{
phyloseq.palmid.help.query(
  srr,
  con = SerratusConnect(),
  metadata = c("run", "bio_project", "bio_sample", "spots", "bases", "tax_id",
    "scientific_name"),
  qc_filter = T
)
}
\arguments{
\item{srr}{character vector of SRA run IDs to generate phyloseq object.}

\item{con}{a connection to the Serratus database. use SerratusConnect by default}

\item{metadata}{character vector of sraruns table columns that should be extracted for each SRR sample}
}
\value{
4 filtered Serratus tables: 
palm = palm_id counts (coverage) for each SRR sample. Derived from Serratus table palm_sra2.
sra = SRR sample metadata. Derived from Serratus table srarun. 
tax = contains tax_id for each palm_id. Derived from Serratus table palm_gb.
tax_ranks = maps tax_id to taxonomic rank. Derived from Serratus table tax_lineage.
}
\description{
Takes an input of SRR run ids. Returns the 4 data tables from Serratus required to generate phyloseq object. Helper function for phyloseq.palmid.
}
\seealso{
[SerratusConnect()]
}
