% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.palmid.R
\name{phyloseq.palmid}
\alias{phyloseq.palmid}
\title{phyloseq.palmid}
\usage{
phyloseq.palmid(
  srr,
  con = SerratusConnect(),
  metadata = c("run", "bio_project", "bio_sample", "spots", "bases", "tax_id",
    "scientific_name"),
  sample = "run_id",
  taxa = "sotu",
  qc_filter = T,
  threads = 1
)
}
\arguments{
\item{srr}{character vector of SRA run IDs to generate phyloseq object.}

\item{con}{Serratus SQL database connection}

\item{metadata}{list of metadata columns to extract from Serratus table sraruns. Reducing number of columns can decrease query time.}

\item{sample}{sample column to use from palm_sra2 table as character string. Recommend to use either run_id (each run as sample) or Biosample (each biological sample, including replicates, as sample) **WARNING** currently only tested with run_id.}

\item{taxa}{one of c("sotu", "palm_id"). **WARNING** currently only implemented/tested with palm_id}
}
\value{
Phyloseq object containig all samples that can be found in the palm_sra2 Serratus table
}
\description{
Create a phyloseq object for all viruses (palm_id) detected in given SRR runs.
}
