% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.summary.R
\name{phyloseq.summary}
\alias{phyloseq.summary}
\title{phyloseq.summary}
\usage{
phyloseq.summary(dat)
}
\arguments{
\item{dat}{phyloseq object derived from physeq.palmid(). taxa are rows, samples are columns}
}
\value{
Two lists:  
viral_summary  
- taxa_told_detected = total number of counts detected for virus across all samples (same units as otu_table(dat)) 
- taxa_n_samples = count of number of samples a taxa is detected within
- taxa_prevalence = count of number of samples taxa is detected withim, divided by all samples in the study (taxa_n_samples/# sample in phyloseq)

sample_summary  
- total_detected = total viral counts detected within sample (same units as otu_sable(dat))
- viral_community_n = total number of unique taxa detected per sample
}
\description{
Get common summary statistics per OTU and per sample from phyloseq object. Expects metadata derived from physeq.palmid().
}
\seealso{
[phyloseq.normalize(), phyloseq.palmid()]
}
