% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeo2.R
\name{PlotGeo2}
\alias{PlotGeo2}
\title{Create a rich plotly geo map from a palm.sra data.frame}
\usage{
PlotGeo2(palm.sra)
}
\arguments{
\item{palm.sra}{data.frame, created with get.palmSra(pro.df, con)}
}
\value{
A plotly map
}
\description{
Create a rich plotly geo map from a palm.sra data.frame
}
\examples{
# Waxsystermes example data
data("waxsys.palm.sra")
geoSRA <- PlotGeo2( waxsys.palm.sra )

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
