% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGeoReport.R
\name{PlotGeoReport}
\alias{PlotGeoReport}
\title{A multi-plot wrapper to convert a list of SRA 'run_ids' into
a geographic world-map and timeline.}
\usage{
PlotGeoReport(run_ids, con = SerratusConnect())
}
\arguments{
\item{run_ids}{character, vector of SRA run_id}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
A grid-table object. Dimension standard is 800 x 600 px.
}
\description{
A multi-plot wrapper to convert a list of SRA 'run_ids' into
a geographic world-map and timeline.
}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
