% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotID2.R
\name{PlotID2}
\alias{PlotID2}
\title{Plot Percent-identity vs. GenBank-identity of a pro file}
\usage{
PlotID2(pro, html = TRUE)
}
\arguments{
\item{pro}{data.frame, A diamond-aligned pro file}

\item{html}{boolean, include additional parsing for htmlwidget display [TRUE]}
}
\value{
A scatterplot as a ggplot2 object
}
\description{
Plot Percent-identity vs. GenBank-identity of a pro file
}
\examples{
data("waxsys.pro.df")
PlotID(waxsys.pro.df)

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
