% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOrgn.R
\name{PlotOrgn}
\alias{PlotOrgn}
\title{Plot a wordcloud of the}
\usage{
PlotOrgn(orgns)
}
\arguments{
\item{orgns}{character, vector of "scientific_name" from sra run table}
}
\value{
A ggwordcloud object of the "ntop" frequent terms
}
\description{
Plot a wordcloud of the
}
\examples{

# Retrive organism identifiers from SRA Run Info Table
palm.orgn <- get.sraOrgn(run_ids, con)

# Create wordcloud of organism terms
PlotOrgn( palm.orgn )

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
