% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhyBar.R
\name{PlotPhyBar}
\alias{PlotPhyBar}
\title{A multi-plot wrapper to convert a list of SRA 'run_ids' into
a geographic world-map and timeline.}
\usage{
PlotPhyBar(tree.df, p)
}
\arguments{
\item{tree.df}{data.frame, merged tree labels and pro. use get.proPhy()}

\item{p}{phylogeny visualization object. use PlotPhy}
}
\value{
A grid-table object. Dimension standard is 800 x 600 px.
}
\description{
A multi-plot wrapper to convert a list of SRA 'run_ids' into
a geographic world-map and timeline.
}
\examples{
data("waxsys.pro.df")

proTax <- PlotTaxReport( waxsys.pro.df )

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
