% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhyReport.R
\name{PlotPhyReport}
\alias{PlotPhyReport}
\title{A multi-plot wrapper to convert a list of SRA 'run_ids' into
a geographic world-map and timeline.}
\usage{
PlotPhyReport(input.msa, tree.df, tree.phy)
}
\arguments{
\item{pro.df}{data.frame, imported diamond pro df. use get.pro()}
}
\value{
A grid-table object. Dimension standard is 800 x 600 px.
}
\description{
A multi-plot wrapper to convert a list of SRA 'run_ids' into
a geographic world-map and timeline.
}
\examples{
data("waxsys.pro.df")

proTax <- PlotTaxReport( waxsys.pro.df )

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
