% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPP.R
\name{PlotPP}
\alias{PlotPP}
\title{Plot the palmprint-diagram for a palmscan df object}
\usage{
PlotPP(ps)
}
\arguments{
\item{ps}{A palmscan data.frame containing one palmprint}
}
\value{
A gene-diagram as a ggplot2 object
}
\description{
Plot the palmprint-diagram for a palmscan df object
}
\examples{
pp.diagram <- PlotPP(ps)

}
\keyword{palmid}
\keyword{plot}
