% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPro.R
\name{PlotPro}
\alias{PlotPro}
\title{Create PlotID and PlotTax grid-plot}
\usage{
PlotPro(pro)
}
\arguments{
\item{pro}{A diamond-aligned pro file}
}
\value{
A grid-table object. Dimension standard is 800 x 400 px.
}
\description{
Create PlotID and PlotTax grid-plot
}
\examples{
proPlot <- PlotPro(pro)
plot(proPlot)

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
