% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProReport.R
\name{PlotProReport}
\alias{PlotProReport}
\title{Create PlotID and PlotTax grid-plot}
\usage{
PlotProReport(pro, html = FALSE)
}
\arguments{
\item{pro}{data.frame, pro.df object}

\item{html}{boolean, generate htmlWidget instead of ggplot [FALSE]}
}
\value{
A grid-table object. Dimension standard is 800 x 400 px.
}
\description{
Create PlotID and PlotTax grid-plot
}
\examples{
data("waxsys.pro.df")

proPlot <- PlotProReport(waxsys.pro.df)

plot(proPlot)

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
