% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotReport.R
\name{PlotReport}
\alias{PlotReport}
\title{A wrapper for palmid Plot* functions to create a standard "report"}
\usage{
PlotReport(pp, pp.bg)
}
\arguments{
\item{pp}{A palmprint.df row to use for the plot "value"}

\item{pp.bg}{A multiple palmprint.df  for the plot background "distribution"}
}
\value{
A grid-table object. Dimension standard is 800 x 400 px.
}
\description{
A wrapper for palmid Plot* functions to create a standard "report"
}
\examples{
ppRep <- PlotReport(pp, pp.bg)
plot(ppRep)
}
\keyword{fev}
\keyword{palmid}
