% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSraExpr.R
\name{PlotSraExpr}
\alias{PlotSraExpr}
\title{Plot the expression of each palm_id-sra (upto top 10)}
\usage{
PlotSraExpr(palm.sra = NULL, ntop = 10)
}
\arguments{
\item{palm.sra}{data.frame, created from get.palmSra() [NULL]}

\item{ntop}{numeric, number of sOTU to return in plot [10]}
}
\value{
A ggplot of per-sOTU expression
}
\description{
Plot the expression of each palm_id-sra (upto top 10)
}
\examples{

# Load Waxsystermes Example data
data("waxsys.palm.sra")

Plot the contig-coverage of the top10 related sOTU
PlotSraExpr( waxsys.palm.sra )


}
\keyword{expression}
\keyword{palmid}
\keyword{plot}
\keyword{sra}
