% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSTAT.R
\name{PlotSTAT}
\alias{PlotSTAT}
\title{Plot a wordcloud of the STAT-classified taxonomy in a palm.sra object
using get.sraSTAT()}
\usage{
PlotSTAT(palm.sra, stat.sra, freq = FALSE)
}
\arguments{
\item{palm.sra}{data.frame, created from get.palmSra()}

\item{stat.sra}{data.frame, created from get.sraTax()}

\item{freq}{boolean, scale words by frequency, else by identity [FALSE]}
}
\value{
A ggwordcloud object of the "ntop" frequent terms
}
\description{
Plot a wordcloud of the STAT-classified taxonomy in a palm.sra object
using get.sraSTAT()
}
\examples{

# Load Waxsystermes Example data
data("waxsys.palm.sra")

# Create wordcloud of organism terms
# using column "order_name" in data.frame

# Scaled by frequency of organism term in all of data.frame
PlotSTAT( waxsys.palm.sra )

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
