% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTax.R
\name{PlotTax}
\alias{PlotTax}
\title{Plot a taxonomic-classifier based histogram}
\usage{
PlotTax(pro)
}
\arguments{
\item{pro}{A diamond-aligned pro file}
}
\value{
A histogram ggplot2
}
\description{
Plot a taxonomic-classifier based histogram
}
\examples{
proTax <- PlotTax(pro)

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
