% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeline.R
\name{PlotTimeline}
\alias{PlotTimeline}
\title{PlotTimeline
Create a timeline of}
\usage{
PlotTimeline(run_ids = NULL, sra.dates = NULL, con = SerratusConnect())
}
\arguments{
\item{run_ids}{character, vector of sra run_ids}

\item{sra.dates}{POSIXct, set of dates}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
ggplot2, timeline of SRA load dates
}
\description{
PlotTimeline
Create a timeline of
}
\keyword{palmid}
\keyword{sql}
\keyword{timeline}
