% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProjExpr.R
\name{ProjZscore}
\alias{ProjZscore}
\title{Plot the expression of each palm_id-sra (upto top 10)
but grouped by BioProject to look for intra-experiment outliers}
\usage{
ProjZscore(expr.df = NULL)
}
\arguments{
\item{palm.sra}{data.frame, created from get.palmSra() [NULL]}

\item{ntop}{numeric, number of sOTU to return in plot [10]}
}
\value{
A ggplot of per-sOTU expression
}
\description{
Plot the expression of each palm_id-sra (upto top 10)
but grouped by BioProject to look for intra-experiment outliers
}
\examples{

# Load Waxsystermes Example data
data("waxsys.palm.sra")

Plot the contig-coverage of the top10 related sOTU
PlotSraExpr( waxsys.palm.sra )


}
\keyword{expression}
\keyword{palmid}
\keyword{plot}
\keyword{sra}
