% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.sparse.R
\name{rbind.sparse}
\alias{rbind.sparse}
\title{rbind.sparse}
\usage{
\method{rbind}{sparse}(x, y)
}
\arguments{
\item{list}{Two sparse matrices to bind}
}
\value{
Single sparse matrix derived from all row-binded sparse matrices in list.
}
\description{
Bind rows of 2 sparse matrices. Add names.
}
