% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dtb.R
\name{read.dtb}
\alias{read.dtb}
\title{Reads a .dtb.hmmscan file created by 'HMMER' 'hmmscan'}
\usage{
read.dtb(dtb.path)
}
\arguments{
\item{dtb.path}{relative system path to .dtb.hmmscan file}
}
\value{
A hmmer data-frame
}
\description{
Reads a .dtb.hmmscan file created by 'HMMER' 'hmmscan'
}
\examples{

}
\keyword{fev}
\keyword{palmid}
