% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.phy.R
\name{read.phy}
\alias{read.phy}
\title{Reads a .phy file generated from muscle}
\usage{
read.phy(phy.path)
}
\arguments{
\item{phy.path}{relative system path to .fev file}
}
\value{
A phylo object
}
\description{
Reads a .phy file generated from muscle
}
\examples{
# palmDB Tree file (.phy)
phy.path <- system.file( "extdata", "waxsys.phy", package = "palmid")
tree.phy <- read.phy(phy.path)

}
\keyword{muscle}
\keyword{palmid}
\keyword{phylogeny}
\keyword{tree}
