% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SerratusConnect.R
\name{SerratusConnect}
\alias{SerratusConnect}
\title{SerratusConnect}
\usage{
SerratusConnect()
}
\value{
con PostgreSQLConnection
}
\description{
Connection information for reaching the Serratus postgreSQL database
See also: https://github.com/ababaian/serratus/wiki/SQL-Schema
}
\examples{
con <- SerratusConnect()

}
\keyword{Serratus}
\keyword{Tantalus}
\keyword{palmid}
\keyword{postgres}
