% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.summary.R
\name{text.summary}
\alias{text.summary}
\title{Create a human-readable text summary of palmID}
\usage{
\method{text}{summary}(pp.in, pro.df, palm.sra, palmdb.hits)
}
\arguments{
\item{pp.in}{data.frame, palmprint from read.fev()}

\item{pro.df}{data.frame, DIAMOND alignment to palmDB from read.pro()}

\item{palm.sra}{data.frame, palm-sra object from get.palmSra()}

\item{palmdb.hits}{numeric, original value for all-palmDB hits}
}
\value{
A character-string summary to use with cat()
}
\description{
Create a human-readable text summary of palmID
}
\keyword{diamond}
\keyword{palmid}
\keyword{pro}
