% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waxsys.msa.R
\docType{data}
\name{waxsys.msa}
\alias{waxsys.msa}
\title{waxsys.msa}
\format{
AAStringSet with 11 entries
}
\source{
\href{https://github.com/ababaian/serratus/wiki/Find_novel_viruses_B_palmdb}{Waxsystermes Virus Tutorial}
}
\usage{
data(waxsys.msa)
}
\description{
Waxsystermes Virus example data. MUSCLE-generated multiple sequence
alignment of Waxsystermes example data. File imported with
Biostrings::readAAStringSet("inst/extdata/waxsys.msa.fa")
}
\keyword{datasets}
\keyword{waxsystermes-example}
