% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waxsys.palm.sra.R
\docType{data}
\name{waxsys.palm.sra}
\alias{waxsys.palm.sra}
\title{waxsys.palm.sra}
\format{
data.frame with 4159 obs. of 13 variables
}
\source{
\href{https://github.com/ababaian/serratus/wiki/Find_novel_viruses_B_palmdb}{Waxsystermes Virus Tutorial}
}
\usage{
data(waxsys.palm.sra)
}
\description{
Waxsystermes Virus example data. Sequencing libraries in the Sequence Read
Archive (SRA) for which Waxsys or related viruses have been identified in
the Serratus Database. Including associated sequencing library meta-data.
}
\details{
#' \itemize{
  \item run_id. Sequencing library identifier in the SRA
  \item palm_id. PalmDB unique identifier for RdRP sequence in library
  \item coverage. Read coverage of contig containing RdRP
  \item sOTU. Species-like Operational Taxonomic Unit to which palm_id belong
  \item qseqid. Input query sequence name
  \item pident. Identity between input sequence and respective sOTU
  \item evalue. Expectence value for alignment
  \item sra_sequence. Matching sequence in the sequence library
  \item biosample_id. Corresponding identifier for the biosample database
  \item scientific_name. Provided organism meta-data
  \item date. Load date for sequencing library
  \item lng. Longitude meta-data (if available)
  \item lat. Latitude meta-data (if available)
}
}
\keyword{datasets}
\keyword{waxsystermes-example}
