% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waxsys.tree.phy.R
\docType{data}
\name{waxsys.tree.phy}
\alias{waxsys.tree.phy}
\title{waxsys.tree.phy}
\format{
Phylogenetic tree with 11 tips and 10 internal nodes.
}
\source{
\href{https://github.com/ababaian/serratus/wiki/Find_novel_viruses_B_palmdb}{Waxsystermes Virus Tutorial}
}
\usage{
data(waxsys.tree.phy)
}
\description{
Waxsystermes Virus example data. MUSCLE-generated UPGMA Tree file
for the alignment of Waxsystermes example data. File imported with
ggtree::read.tree("inst/extdata/waxsys.phy")
}
\details{
Tip labels:
 u18590_41.8, u8640_41.4, u181012_43.0, u32314_41.9, u253902_41.8, u38234_41.7, ...

Rooted; includes branch lengths.
}
\keyword{datasets}
\keyword{waxsystermes-example}
