% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoFilter.R
\name{geoFilter}
\alias{geoFilter}
\title{Conversion between run_ids and geo objects often contain NA/NULL values
This removes NA-containing rows}
\usage{
geoFilter(sra.geo, wobble = FALSE, wradius = 0.005)
}
\arguments{
\item{sra.geo}{data.frame, output df from get.sraGeo()}

\item{wobble}{boolean, add a wobble to each point to prevent overplotting [FALSE]}

\item{wradius}{numeric, maximum magnitude of wobble [0.005]}
}
\value{
data.frame
}
\description{
Conversion between run_ids and geo objects often contain NA/NULL values
This removes NA-containing rows
}
\keyword{geo}
\keyword{palmid}
