% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.proPhy.R
\name{get.proPhy}
\alias{get.proPhy}
\title{Merges 'pro.df' with 'tree.phy'
to add annotations from palmDB to the phylogenetic tree object.
Returns a data frame containing tree tip labels and palmDB columns}
\usage{
get.proPhy(pro.df, tree.phy)
}
\arguments{
\item{pro.df}{data.frame, imported diamond pro df. use get.pro()}

\item{tree.phy}{phylo object. use read.phy}
}
\value{
tree.df  data.frame
}
\description{
Merges 'pro.df' with 'tree.phy'
to add annotations from palmDB to the phylogenetic tree object.
Returns a data frame containing tree tip labels and palmDB columns
}
\examples{

## Prepare data
# data("waxsys.pro.df")
# phy.path <- system.file( "extdata", "waxsys.phy", package = "palmid")
# waxsys.phy <- read.phy(phy.path)

## Generate tree data frame
# tree.df <- get.proPhy(waxsys.pro.df, waxsys.phy)


}
\keyword{muscle}
\keyword{palmid}
\keyword{phylogeny}
\keyword{tree}
