% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sra.R
\name{get.sra}
\alias{get.sra}
\title{get.sra}
\usage{
get.sra(palm_ids, con, get_contig.df = FALSE, qc = TRUE)
}
\arguments{
\item{palm_ids}{character, set of `palm_id` to lookup in palmdb}

\item{con}{pq-connection, use SerratusConnect()}

\item{get_contig.df}{boolean, return a data.frame of matching contigs [F]}

\item{qc}{boolean, require 85% palmprint coverage and e-value < 1e-6}
}
\value{
character, de-duplicated run_ids with a potential match for `palm_ids`
}
\description{
Retrieve the SRA runs containing a microassemblied contig matching
a given `palm_id`
}
\examples{
palm.sra   <- get.sra(palm.group, con)

}
\keyword{Serratus}
\keyword{palmdb}
\keyword{palmid}
\keyword{sOTU}
