% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraBio.R
\name{get.sraBio}
\alias{get.sraBio}
\title{get.sraBio}
\usage{
get.sraBio(run_ids, con, ordinal = F)
}
\arguments{
\item{run_ids}{character, SRA `run_id`}

\item{con}{pq-connection, use SerratusConnect()}

\item{ordinal}{boolean, return `run_ids` ordered vector [F]}
}
\value{
data.frame, run_id, biosample character vectors
}
\description{
Retrieve the "BioSample" field for a set of SRA `run_id`
}
\examples{
palm.bio   <- get.sraBio(palm.sras, con)

}
\keyword{Serratus}
\keyword{geo}
\keyword{palmid}
