% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraBio.R, R/get.sraProj.R
\name{get.sraBio}
\alias{get.sraBio}
\title{get.sraBio}
\usage{
get.sraBio(run_ids, con, ordinal = FALSE)

get.sraBio(run_ids, con, ordinal = FALSE)
}
\arguments{
\item{run_ids}{character, SRA 'run_id'}

\item{con}{pq-connection, use SerratusConnect()}

\item{ordinal}{boolean, return 'run_ids' ordered vector [FALSE]}

\item{biodb}{character, one of "both", "BioProject", or ["BioSample"]}
}
\value{
data.frame, run_id, biosample character vectors

data.frame, run_id, biosample character vectors
}
\description{
Retrieve the "BioSample", "BioProject", or "Both" field for a set of 
input SRA 'run_id'

Retrieve the "BioProject" field for a set of SRA 'run_id'
and return all SRA run_id for each unique BioProject
<UNDER DEVELOPMENT>
}
\examples{
\donttest{
# SRA Library of interest
con <- SerratusConnect()
library.bioSample   <- get.sraBio( 'SRR9968562' , con)
}
\donttest{
# SRA Library of interest
con <- SerratusConnect()
library.bioProject   <- get.sraProj( 'SRR9968562' , con)
}
}
\keyword{BioProject}
\keyword{SRA}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{palmid}
\keyword{sra}
