% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraOrgn.R
\name{get.sraOrgn}
\alias{get.sraOrgn}
\title{get.sraOrgn}
\usage{
get.sraOrgn(run_ids, con, as.df = FALSE)
}
\arguments{
\item{run_ids}{character, SRA `run_id`}

\item{con}{pq-connection, use SerratusConnect()}

\item{as.df}{boolean, return run_id, date data.frame [F]}
}
\value{
character, string vector
}
\description{
Retrieve the 'scientific_name' for a set of SRA `run_id`
}
\examples{
palm.orgn   <- get.sraOrg(palm.group, con)

}
\keyword{Serratus}
\keyword{palmid}
\keyword{taxonomy}
