% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.lib.R
\name{load.lib}
\alias{load.lib}
\title{load.lib.R}
\usage{
load.lib(lib_set)
}
\arguments{
\item{lib_set}{character, one of "sql", "ggplot", "geo"}
}
\value{
loads set of packages for functions
}
\description{
Internal function to load a set of libraries for palmid functions
}
\keyword{palmid}
\keyword{serratus}
\keyword{sql}
