% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmdb-data.R
\docType{data}
\name{palmdb}
\alias{palmdb}
\title{palmdb <210302>}
\format{
a multiple-palmprint data.frame
}
\source{
\href{https://github.com/rcedgar/palmdb}{palmDB repo}
}
\usage{
data(palmdb)
}
\description{
Palmprint database - initial version (210302) containing GenBank v241
RNA depenendent RNA polymerase palmprints
}
\examples{
data(palmdb)
}
\keyword{datasets}
\keyword{palmdb}
