% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotID.R
\name{PlotID}
\alias{PlotID}
\title{Plot %-identity vs. E-value of a pro file}
\usage{
PlotID(pro)
}
\arguments{
\item{pro}{A diamond-aligned pro file}
}
\value{
A scatterplot as a ggplot2 object
}
\description{
Plot %-identity vs. E-value of a pro file
}
\examples{
idPlot <- PlotID(pro)

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
