% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhy.R
\name{PlotPhy}
\alias{PlotPhy}
\title{A phylogeny plot wrapper to visualize a phylo object
that contains tree tip labels from top-10 hits from palmDB
and associated metadata from 'pro.df'}
\usage{
PlotPhy(tree.df, tree.phy)
}
\arguments{
\item{tree.df}{data.frame, merged tree labels and pro.df. use get.proPhy}

\item{tree.phy}{phylo object of phylogeny of top-10 hits. use read.phy}
}
\value{
A phylogeny plot with readable tip labels.
}
\description{
A phylogeny plot wrapper to visualize a phylo object
that contains tree tip labels from top-10 hits from palmDB
and associated metadata from 'pro.df'
}
\examples{
data("waxsys.tree.df")
data("waxsys.tree.phy")

p <- PlotPhy(waxsys.tree.df, waxsys.tree.phy)

}
\keyword{dplyr}
\keyword{ggplot2}
\keyword{ggtree}
\keyword{treeio}
