% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhyBar.R
\name{PlotPhyBar}
\alias{PlotPhyBar}
\title{A plot wrapper to generate a horizontal bar plot
of input identity to palmDB (aa%).
The order of the bars matches the taxa order
of the phylogenetic tree object tree.phy}
\usage{
PlotPhyBar(tree.df, p)
}
\arguments{
\item{tree.df}{data.frame, merged tree labels and pro.df. use get.proPhy}

\item{p}{a ggtree plot of tree.phy. use PlotPhy}
}
\value{
A bar plot of % input identity to palmDB
}
\description{
A plot wrapper to generate a horizontal bar plot
of input identity to palmDB (aa%).
The order of the bars matches the taxa order
of the phylogenetic tree object tree.phy
}
\examples{
data("waxsys.tree.df")
data("waxsys.tree.phy")

p <- PlotPhy(waxsys.tree.df, waxsys.tree.phy)
plotPhyBar <- PlotPhyBar(waxsys.tree.df, p)

}
\keyword{dplyr}
\keyword{ggplot2}
\keyword{ggtree}
\keyword{treeio}
