% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPhyMsa.R
\name{PlotPhyMsa}
\alias{PlotPhyMsa}
\title{A multi-plot wrapper to generate an MSA plot from input.msa and
align it with a phylogenetic tree plot.}
\usage{
PlotPhyMsa(input.msa, p)
}
\arguments{
\item{input.msa}{msa file path relative to .fev}

\item{p}{phylogeny visualization object. use PlotPhy}
}
\value{
A single object consisting of two facets for tree and MSA
}
\description{
A multi-plot wrapper to generate an MSA plot from input.msa and
align it with a phylogenetic tree plot.
}
\examples{
data("waxsys.input.msa")
data("waxsys.tree.phy")
p <- PlotPhy(waxsys.tree.df, waxsys.tree.phy)
plotPhyMsa <- PlotPhyMsa(waxsys.input.msa, p)

}
\keyword{dplyr}
\keyword{ggplot2}
\keyword{ggtree}
\keyword{msa}
\keyword{treeio}
