% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSTAT.R
\name{PlotSTAT}
\alias{PlotSTAT}
\title{Plot a wordcloud of the STAT-classified taxonomy in a palm.sra object
using get.sraSTAT()}
\usage{
PlotSTAT(stat.sra)
}
\arguments{
\item{stat.sra}{data.frame, created from get.sraSTAT()}
}
\value{
A ggwordcloud object of the "ntop" frequent terms [50]
}
\description{
Plot a wordcloud of the STAT-classified taxonomy in a palm.sra object
using get.sraSTAT()
}
\examples{
\donttest{

con <- SerratusConnect()

# Load Waxsystermes Example data
data("waxsys.palm.sra")

waxsys.stat.sra <- get.sraSTAT(waxsys.palm.sra$run_id, con)

# Populate stat.sra with percent identity from palm.sra
palm.stat.lookup <- match(waxsys.stat.sra$run_id, waxsys.palm.sra$run_id)
waxsys.stat.sra$pident <- waxsys.palm.sra$pident[palm.stat.lookup]

# Create wordcloud of STAT Taxonomy
PlotSTAT( waxsys.stat.sra)

}

}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
