% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTax.R
\name{PlotTax}
\alias{PlotTax}
\title{Plot a taxonomic-classifier based histogram}
\usage{
PlotTax(pro, html = TRUE)
}
\arguments{
\item{pro}{data.frame, A diamond-aligned pro file}

\item{html}{boolean, include additional parsing for htmlwidget display [TRUE]}
}
\value{
A histogram ggplot2
}
\description{
Plot a taxonomic-classifier based histogram
}
\examples{
data("waxsys.pro.df")

proTax <- PlotTax(waxsys.pro.df, html = TRUE)

plot(proTax)
}
\keyword{palmid}
\keyword{plot}
\keyword{pro}
