% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeline.R
\name{PlotTimeline}
\alias{PlotTimeline}
\title{PlotTimeline
Create a timeline of data-releases in palm.sra}
\usage{
PlotTimeline(palm.sra, html = TRUE)
}
\arguments{
\item{palm.sra}{data.frame, created with get.palmSra(pro.df, con)}

\item{html}{boolean, generate htmlWidget instead of ggplot [TRUE]}
}
\value{
ggplot2, timeline of SRA load dates
}
\description{
PlotTimeline
Create a timeline of data-releases in palm.sra
}
\keyword{palmid}
\keyword{timeline}
