% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTimeline.R
\name{PlotTimeline}
\alias{PlotTimeline}
\title{PlotTimeline
Create a timeline of}
\usage{
PlotTimeline(sra.date)
}
\arguments{
\item{sra.date}{POSIXct, set of dates}
}
\value{
ggplot2,
}
\description{
PlotTimeline
Create a timeline of
}
\examples{
sc2.timeline <- PlotTimeline(sra.date)
plot(sc2.timeline)

}
\keyword{palmid}
\keyword{timeline}
