% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.phy.R
\name{read.phy}
\alias{read.phy}
\title{Reads a .pro file created by 'diamond'}
\usage{
read.phy(input.phy)
}
\arguments{
\item{pro.path}{relative system path to .fev file}
}
\value{
A diamond-pro data.frame object
}
\description{
Reads a .pro file created by 'diamond'
}
\examples{
# palmDB Alignment file (.pro)
pro.path <- system.file( "extdata", "waxsys.pro", package = "palmid")
pro.df <- read.pro(pro.path)

}
\keyword{diamond}
\keyword{palmid}
\keyword{pro}
