% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeWordCount.R
\name{standardizeWordcount}
\alias{standardizeWordcount}
\title{standardizeWordcount
Create frequency-count table from a set of characters
which are assigned a standardized rank-order scores
from 10.0 to 1.0.}
\usage{
standardizeWordcount(words, ntop = 50)
}
\arguments{
\item{words}{character, description of vector [Default]}

\item{ntop}{numeric,   return only N top words [50]}
}
\value{
table, frequency-count table with 10:1 standard
}
\description{
standardizeWordcount
Create frequency-count table from a set of characters
which are assigned a standardized rank-order scores
from 10.0 to 1.0.
}
\examples{
# Desc
function_name( var1 = x, var2 = T)

}
\keyword{palmid}
\keyword{plot}
\keyword{wordcloud}
